+++ Content/Enemies/Bosses/Grovetender.cs
@@ -14,6 +14,28 @@
 
         public override void Hooks()
         {
+            On.EntityStates.GravekeeperBoss.PrepHook.OnEnter += PrepHook_OnEnter;
+            On.EntityStates.GravekeeperBoss.FireHook.OnEnter += FireHook_OnEnter;
+            Changes();
+        }
+
+        public static GameObject hookPrefab = Utils.Paths.GameObject.GravekeeperHookProjectile.Load<GameObject>();
+
+        private void FireHook_OnEnter(On.EntityStates.GravekeeperBoss.FireHook.orig_OnEnter orig, EntityStates.GravekeeperBoss.FireHook self)
+        {
+            EntityStates.GravekeeperBoss.FireHook.projectilePrefab = hookPrefab;
+            EntityStates.GravekeeperBoss.FireHook.projectileForce = -800f;
+            orig(self);
+        }
+
+        private void PrepHook_OnEnter(On.EntityStates.GravekeeperBoss.PrepHook.orig_OnEnter orig, EntityStates.GravekeeperBoss.PrepHook self)
+        {
+            EntityStates.GravekeeperBoss.PrepHook.baseDuration = 1.3f;
+            orig(self);
+        }
+
+        private void Changes()
+        {
             if (shouldReplaceClayDunestriderOnSunderedGrove)
             {
                 var sunderedGrove = Utils.Paths.DirectorCardCategorySelection.dccsRootJungleMonsters.Load<DirectorCardCategorySelection>();
@@ -22,6 +44,10 @@
                 var sunderedGroveDLC1 = Utils.Paths.DirectorCardCategorySelection.dccsRootJungleMonstersDLC1.Load<DirectorCardCategorySelection>();
                 sunderedGroveDLC1.categories[0] /* champions */.cards[0] /* clay dunestrider */.spawnCard = Utils.Paths.CharacterSpawnCard.cscGravekeeper.Load<CharacterSpawnCard>();
             }
+
+            var wisp = Utils.Paths.GameObject.GravekeeperTrackingFireball.Load<GameObject>().GetComponent<CharacterBody>();
+            wisp.baseMaxHealth = 35f;
+            wisp.levelMaxHealth = 10.5f;
         }
     }
 }