+++ Content/Enemies/Bosses/StoneTitan.cs
@@ -112,8 +112,7 @@ namespace WellRoundedBalance.Enemies.Bosses
 
         private void Changes()
         {
-            ContentAddition.AddEntityState(typeof(ChargeLaserSpam), out _);
-            ContentAddition.AddEntityState(typeof(FireLaserSpam), out _);
+            ContentAddition.AddEntityState(typeof(LaserAttack), out _);
 
             var rockController = Utils.Paths.GameObject.TitanRockController.Load<GameObject>();
             var destroyOnTimer = rockController.GetComponent<DestroyOnTimer>();
@@ -139,7 +138,8 @@ namespace WellRoundedBalance.Enemies.Bosses
         }
     }
 
-    public class LaserAttack : BaseState {
+    public class LaserAttack : BaseState
+    {
         public float durationPerLaser => 0.4f;
         public int baseLasers => 12;
         public static GameObject effect = Utils.Paths.GameObject.LaserGolem.Load<GameObject>();
@@ -160,10 +160,10 @@ namespace WellRoundedBalance.Enemies.Bosses
         {
             base.OnEnter();
 
-            laserInstance = GameObject.Instantiate(effect, base.FindModelChild("MuzzleLaser"));
+            laserInstance = Object.Instantiate(effect, FindModelChild("MuzzleLaser"));
             lr = laserInstance.GetComponent<LineRenderer>();
 
-            chargeID = AkSoundEngine.PostEvent(Events.Play_titanboss_R_laser_preshoot, base.gameObject);
+            chargeID = AkSoundEngine.PostEvent(Events.Play_titanboss_R_laser_preshoot, gameObject);
         }
 
         public override void OnExit()
@@ -179,7 +179,7 @@ namespace WellRoundedBalance.Enemies.Bosses
 
             stopwatch += Time.fixedDeltaTime;
 
-            Ray aimRay = base.GetAimRay();
+            Ray aimRay = GetAimRay();
             float width = durationPerLaser - stopwatch < 0.1f ? 0.1f : durationPerLaser - stopwatch;
 
             lr.startWidth = width;
@@ -189,299 +189,47 @@ namespace WellRoundedBalance.Enemies.Bosses
 
             laserInstance.transform.forward = targetDir;
 
-            if (durationPerLaser - stopwatch > durationPerLaser / 2) {
+            if (durationPerLaser - stopwatch > durationPerLaser / 2)
+            {
                 targetDir = aimRay.direction; // grace period before firing so it's dodgeable
             }
 
-            if (stopwatch >= delay) {
+            if (stopwatch >= delay)
+            {
                 count++;
                 stopwatch = 0f;
 
-                BulletAttack attack = new();
-                attack.origin = aimRay.origin;
-                attack.damage = base.damageStat * 2f;
-                attack.procCoefficient = 0.6f;
-                attack.falloffModel = BulletAttack.FalloffModel.None;
-                attack.aimVector = targetDir;
-                attack.muzzleName = "MuzzleLaser";
-                attack.tracerEffectPrefab = tracer;
-                attack.minSpread = 0;
-                attack.maxSpread = 0;
-                attack.maxDistance = 2000;
-                attack.owner = base.gameObject;
-                attack.smartCollision = true;
-                attack.radius = 1f;
-                attack.force = 400f;
-                attack.queryTriggerInteraction = QueryTriggerInteraction.Ignore;
+                BulletAttack attack = new()
+                {
+                    origin = aimRay.origin,
+                    damage = damageStat,
+                    procCoefficient = 0.6f,
+                    falloffModel = BulletAttack.FalloffModel.None,
+                    aimVector = targetDir,
+                    muzzleName = "MuzzleLaser",
+                    tracerEffectPrefab = tracer,
+                    minSpread = 0,
+                    maxSpread = 0,
+                    maxDistance = 2000,
+                    owner = gameObject,
+                    smartCollision = true,
+                    radius = 1f,
+                    force = 400f,
+                    queryTriggerInteraction = QueryTriggerInteraction.Ignore
+                };
 
-                if (base.isAuthority) {
+                if (base.isAuthority)
+                {
                     attack.Fire();
                 }
 
                 AkSoundEngine.PostEvent(Events.Play_golem_laser_fire, base.gameObject);
 
-                if (count > baseLasers) {
+                if (count > baseLasers)
+                {
                     outer.SetNextStateToMain();
                 }
             }
-        }
-    }
-
-    public class ChargeLaserSpam : BaseState
-    {
-        public static float baseDuration = 3f;
-
-        public static float laserMaxWidth = 0.2f;
-
-        public static GameObject effectPrefab = Utils.Paths.GameObject.ChargeGolem.Load<GameObject>();
-
-        public static GameObject laserPrefab = Utils.Paths.GameObject.LaserGolem.Load<GameObject>();
-
-        public static string attackSoundString = "Play_titanboss_R_laser_preshoot";
-
-        public float duration;
-
-        public uint chargePlayID;
-
-        public GameObject chargeEffect;
-
-        public GameObject laserEffect;
-
-        public LineRenderer laserLineComponent;
-
-        public Vector3 laserDirection;
-
-        public float flashTimer = 0f;
-
-        public bool laserOn = true;
-
-        public override void OnEnter()
-        {
-            base.OnEnter();
-            duration = baseDuration / attackSpeedStat;
-            var modelTransform = GetModelTransform();
-            chargePlayID = Util.PlayAttackSpeedSound(attackSoundString, gameObject, attackSpeedStat);
-            if (modelTransform)
-            {
-                var childLocator = modelTransform.GetComponent<ChildLocator>();
-                if (childLocator)
-                {
-                    var laserMuzzle = childLocator.FindChild("MuzzleLaser");
-                    if (laserMuzzle)
-                    {
-                        if (effectPrefab)
-                        {
-                            chargeEffect = Object.Instantiate(effectPrefab, laserMuzzle.position, laserMuzzle.rotation);
-                            chargeEffect.transform.parent = laserMuzzle;
-                            var scaleParticleSystemDuration = chargeEffect.GetComponent<ScaleParticleSystemDuration>();
-                            if (scaleParticleSystemDuration)
-                            {
-                                scaleParticleSystemDuration.newDuration = duration;
-                            }
-                        }
-                        if (laserPrefab)
-                        {
-                            laserEffect = Object.Instantiate(laserPrefab, laserMuzzle.position, laserMuzzle.rotation);
-                            laserEffect.transform.parent = laserMuzzle;
-                            laserLineComponent = laserEffect.GetComponent<LineRenderer>();
-                        }
-                    }
-                }
-            }
-            if (characterBody)
-            {
-                // characterBody.SetAimTimer(duration);
-            }
-            flashTimer = 0f;
-            laserOn = true;
-        }
-
-        public override void OnExit()
-        {
-            AkSoundEngine.StopPlayingID(chargePlayID);
-            base.OnExit();
-            if (chargeEffect)
-            {
-                Destroy(chargeEffect);
-            }
-            if (laserEffect)
-            {
-                Destroy(laserEffect);
-            }
-        }
-
-        public override void Update()
-        {
-            base.Update();
-            if (laserEffect && laserLineComponent)
-            {
-                var aimRay = GetAimRay();
-                var position = laserEffect.transform.parent.position;
-                var endPoint = aimRay.GetPoint(1000f);
-                // laserDirection = endPoint - position;
-                laserDirection = aimRay.direction;
-                if (Physics.Raycast(aimRay, out RaycastHit raycastHit, 1000f, LayerIndex.world.mask | LayerIndex.entityPrecise.mask))
-                {
-                    endPoint = raycastHit.point;
-                }
-                laserLineComponent.SetPosition(0, position);
-                laserLineComponent.SetPosition(1, endPoint);
-                float totalWidth;
-                if (duration - age > 0.5f)
-                {
-                    totalWidth = age / duration;
-                }
-                else
-                {
-                    flashTimer -= Time.deltaTime;
-                    if (flashTimer <= 0f)
-                    {
-                        laserOn = !laserOn;
-                        flashTimer = 0.033333335f;
-                    }
-                    totalWidth = (laserOn ? 1f : 0f);
-                }
-                totalWidth *= laserMaxWidth;
-                laserLineComponent.startWidth = totalWidth;
-                laserLineComponent.endWidth = totalWidth;
-            }
-        }
-
-        public override void FixedUpdate()
-        {
-            base.FixedUpdate();
-            if (fixedAge >= duration && isAuthority)
-            {
-                FireLaserSpam fireLaserSpam = new()
-                {
-                    laserDirection = laserDirection
-                };
-                outer.SetNextState(fireLaserSpam);
-                return;
-            }
-        }
-
-        public override InterruptPriority GetMinimumInterruptPriority()
-        {
-            return InterruptPriority.Skill;
-        }
-    }
-
-    public class FireLaserSpam : BaseState
-    {
-        public static GameObject effectPrefab = Utils.Paths.GameObject.MuzzleflashGolem.Load<GameObject>();
-
-        public static GameObject hitEffectPrefab = Utils.Paths.GameObject.ExplosionGolem.Load<GameObject>();
-
-        public static GameObject tracerEffectPrefab = Utils.Paths.GameObject.TracerGolem.Load<GameObject>();
-
-        public static float damageCoefficient = 1f;
-
-        public static float blastRadius = 3f;
-
-        public static float force = 1500f;
-
-        public static float minSpread = 0f;
-
-        public static float maxSpread = 1f;
-
-        public static float maxBaseDuration = 5f;
-        public static float baseDuration = 0.5f;
-
-        public static string attackSoundString = "Play_golem_laser_fire";
-
-        public Vector3 laserDirection;
-
-        public float duration;
-
-        public Ray modifiedAimRay;
-        public float timer;
-
-        public override void OnEnter()
-        {
-            base.OnEnter();
-            duration = baseDuration / attackSpeedStat;
-            PlayCrossfade("Gesture, Additive", "FireLaserLoop", 0.25f);
-            if (characterBody)
-            {
-                characterBody.SetAimTimer(maxBaseDuration);
-            }
-        }
-
-        public override void OnExit()
-        {
-            base.OnExit();
-            PlayCrossfade("Gesture, Additive", "FireLaserEnd", 0.25f);
-        }
-
-        public override void FixedUpdate()
-        {
-            base.FixedUpdate();
-            timer += Time.fixedDeltaTime;
-            if (timer >= duration)
-            {
-                FireLaser();
-                timer = 0f;
-            }
-            if (fixedAge >= maxBaseDuration && isAuthority)
-            {
-                outer.SetNextStateToMain();
-                return;
-            }
-        }
-
-        public void FireLaser()
-        {
-            modifiedAimRay = GetAimRay();
-            modifiedAimRay.direction = laserDirection;
-            GetModelAnimator();
-            Util.PlaySound(attackSoundString, gameObject);
-            var laserMuzzle = "MuzzleLaser";
-            PlayAnimation("Gesture", "FireLaser", "FireLaser.playbackRate", duration);
-            EffectManager.SimpleMuzzleFlash(effectPrefab, gameObject, laserMuzzle, false);
-            var modelTransform = GetModelTransform();
-            if (isAuthority)
-            {
-                float num = 1000f;
-                Vector3 vector = modifiedAimRay.origin + modifiedAimRay.direction * num;
-                if (Physics.Raycast(modifiedAimRay, out RaycastHit raycastHit, num, LayerIndex.world.mask | LayerIndex.defaultLayer.mask | LayerIndex.entityPrecise.mask))
-                {
-                    vector = raycastHit.point;
-                }
-                new BlastAttack
-                {
-                    attacker = gameObject,
-                    inflictor = gameObject,
-                    teamIndex = TeamComponent.GetObjectTeam(gameObject),
-                    baseDamage = damageStat * damageCoefficient,
-                    baseForce = force * 0.2f,
-                    position = vector,
-                    radius = blastRadius,
-                    falloffModel = BlastAttack.FalloffModel.SweetSpot,
-                    bonusForce = force * modifiedAimRay.direction
-                }.Fire();
-                if (modelTransform)
-                {
-                    var childLocator = modelTransform.GetComponent<ChildLocator>();
-                    if (childLocator)
-                    {
-                        var laserMuzzleIndex = childLocator.FindChildIndex(laserMuzzle);
-                        EffectData effectData = new()
-                        {
-                            origin = vector,
-                            start = modifiedAimRay.origin
-                        };
-                        effectData.SetChildLocatorTransformReference(gameObject, laserMuzzleIndex);
-                        EffectManager.SpawnEffect(tracerEffectPrefab, effectData, true);
-                        EffectManager.SpawnEffect(hitEffectPrefab, effectData, true);
-                    }
-                }
-            }
-        }
-
-        public override InterruptPriority GetMinimumInterruptPriority()
-        {
-            return InterruptPriority.Skill;
         }
     }
 }