+++ Content/Enemies/Standard/Geep.cs
@@ -1,4 +1,7 @@
-﻿namespace WellRoundedBalance.Enemies.Standard
+﻿using RoR2.Skills;
+using System;
+
+namespace WellRoundedBalance.Enemies.Standard
 {
     internal class Geep : EnemyBase<Geep>
     {
@@ -15,6 +18,7 @@
         public override void Hooks()
         {
             CharacterBody.onBodyStartGlobal += CharacterBody_onBodyStartGlobal;
+            Changes();
         }
 
         private void CharacterBody_onBodyStartGlobal(CharacterBody body)
@@ -33,11 +37,30 @@
 
         private void Changes()
         {
-            var geep = Addressables.LoadAssetAsync<GameObject>("RoR2/DLC1/Gup/GeepBody.prefab").WaitForCompletion();
+            var geep = Utils.Paths.GameObject.GeepBody12.Load<GameObject>();
 
             var geepBody = geep.GetComponent<CharacterBody>();
             geepBody.baseMaxHealth = baseMaxHealth;
             geepBody.levelMaxHealth = baseMaxHealth * 0.3f;
+
+            var modelTransform = geep.transform.GetChild(0).GetChild(0);
+            var spikes = Array.Find(modelTransform.GetComponents<HitBoxGroup>(), (HitBoxGroup element) => element.groupName == "Spikes");
+            var mainHitbox = spikes.hitBoxes[0].gameObject;
+            mainHitbox.transform.localScale = new Vector3(6f, 6f, 2.5f);
+            /*
+            var hitboxGroup = modelTransform.GetComponent<HitBoxGroup>();
+            hitboxGroup.hitBoxes[0] = spikes.hitBoxes[1];
+            var newHitbox = hitboxGroup.hitBoxes[0];
+            newHitbox.transform.localScale = new Vector3(2f, 2f, 1.4f);
+
+            var contactDamage = geep.AddComponent<ContactDamage>();
+            contactDamage.pushForcePerSecond = 500f;
+            contactDamage.damagePerSecondCoefficient = 0.7f;
+            contactDamage.hitBoxGroup = hitboxGroup;
+            */
+
+            var sd = Utils.Paths.SkillDef.GupSpikes.Load<SkillDef>();
+            sd.baseRechargeInterval = 1.5f;
         }
     }
 }