+++ Content/Mechanics/Monsters/Frog.cs
@@ -6,6 +6,9 @@
         [ConfigField("Max Pets", 1)]
         public static int FrogPets;
 
+        [ConfigField("Max Pets", "", 1)]
+        public static int maxPets;
+
         public override void Init()
         {
             base.Init();
@@ -18,7 +21,7 @@
 
         private void FrogController_Pet(On.RoR2.FrogController.orig_Pet orig, FrogController self, Interactor interactor)
         {
-            self.maxPets = 1;
+            self.maxPets = maxPets;
             orig(self, interactor);
         }
     }